Refactor request-checking to core mixin
